refactoring out default dates and transform functions
